requirements.txt
pandas == 1.3.5
numpy ==1.21.5
matplotlib == 3.5.1
scikit-learn == 1.0.1
tensorflow-gpu == 2.7.0
seaborn == 0.11.2
neurolab ==0.3.5

